Move from scss to regular css with vite
