+++ package.json
@@ -31,7 +31,6 @@
     "classnames": "^2.2.6",
     "debounce-promise": "^3.1.2",
     "formik": "^2.1.5",
-    "normalize-scss": "^7.0.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
@@ -51,7 +50,6 @@
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react-hooks": "^4.0.8",
     "prettier": "^2.0.5",
-    "sass": "^1.80.0",
     "vite": "^5.4.0"
   },
   "browserslist": {
