+++ src/components/Browse/Browse.module.css
@@ -1,4 +1,6 @@
-@import '../../global-styles/base.module.scss';
+.pageTitle {
+  composes: pageTitle from '../../global-styles/base.module.css';
+}
 
 .browseList {
   display: grid;
@@ -25,7 +27,7 @@
   cursor: pointer;
 
   &:hover {
-    svg {
+    & svg {
       transition: all 0.2s ease;
       transform: scale(1.2);
     }
