+++ src/components/Browse/CategoryList.module.css
@@ -33,7 +33,7 @@
   cursor: pointer;
 
   &:hover {
-    svg {
+    & svg {
       transition: all 0.2s ease;
       transform: scale(1.2);
     }
