+++ src/components/HomePage/Homepage.module.css
@@ -1,5 +1,3 @@
-@import '../../global-styles/base.module.scss';
-
 .heroImageContainer {
   height: 100vh;
   width: 100%;
@@ -19,7 +17,7 @@
 }
 
 .navButton {
-  composes: transparentButton;
+  composes: transparentButton from '../../global-styles/base.module.css';
   background-color: rgba(25, 25, 25, 0.3);
 }
