+++ src/components/HomePage/TitleSearch.module.css
@@ -1,12 +1,3 @@
-@import '../../global-styles/base.module.scss';
-
-$inputHeight: 40px;
-$resultHeight: 55px;
-$inputWidth: 410px;
-$buttonWidth: 80px;
-$buttonSpacing: 10px;
-$padding: 10px;
-
 .container {
   position: relative;
   display: flex;
@@ -40,9 +31,8 @@ $padding: 10px;
   color: white;
   font-size: 16px;
   font-weight: 400;
-
-  width: $inputWidth;
-  height: $inputHeight;
+  width: 410px;
+  height: 40px;
   outline: none;
 
   &::-webkit-search-cancel-button {
@@ -50,8 +40,7 @@ $padding: 10px;
     height: 1em;
     width: 1em;
     border-radius: 50em;
-    background: url(https://pro.fontawesome.com/releases/v5.10.0/svgs/solid/times-circle.svg)
-      no-repeat 50% 50%;
+    background: url(https://pro.fontawesome.com/releases/v5.10.0/svgs/solid/times-circle.svg) no-repeat 50% 50%;
     background-size: contain;
     opacity: 0.5;
     color: white;
@@ -69,9 +58,9 @@ $padding: 10px;
 }
 
 .submitButton {
-  height: $inputHeight;
-  min-width: $buttonWidth;
-  margin-left: $buttonSpacing;
+  height: 40px;
+  min-width: 80px;
+  margin-left: 10px;
   font-family: Signika;
   padding: 0 10px;
 
@@ -87,7 +76,7 @@ $padding: 10px;
 
   @media screen and (max-width: 735px) {
     margin-top: 10px;
-    width: $inputWidth;
+    width: 410px;
   }
 
   @media screen and (max-width: 415px) {
@@ -96,12 +85,12 @@ $padding: 10px;
 }
 
 .noneFound {
-  padding: $padding;
-  width: $inputWidth + $buttonWidth + $buttonSpacing;
+  padding: 10px;
+  width: 500px;
   color: white;
 
   @media screen and (max-width: 735px) {
-    width: $inputWidth;
+    width: 410px;
   }
 
   @media screen and (max-width: 415px) {
@@ -109,12 +98,11 @@ $padding: 10px;
   }
 }
 
-// Search Results
 .recipeList {
   position: absolute;
   top: 100%;
   z-index: 2;
-  width: $inputWidth;
+  width: 410px;
   background: white;
   list-style-type: none;
   margin: 0;
@@ -134,10 +122,10 @@ $padding: 10px;
 
 .searchResult {
   display: flex;
-  height: $resultHeight;
-  line-height: $resultHeight;
+  height: 55px;
+  line-height: 55px;
   border-top: 1px solid #dedede;
-  padding: 0 $padding;
+  padding: 0 10px;
   cursor: pointer;
 
   &:first-child {
