+++ src/components/RecipeForm/FormComponents.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames';
 import { Field, ErrorMessage } from 'formik';
-import classes from './RecipeForm.module.scss';
+import classes from './RecipeForm.module.css';
 
 type InputFieldProps = {
   labelText?: string;
