+++ src/components/RecipeView/Ingredients.module.css
@@ -1,5 +1,3 @@
-@import '../../global-styles/base.module.scss';
-
 .sectionTitle {
   text-transform: uppercase;
   font-weight: 700;
@@ -32,7 +30,7 @@
   font-size: 0.8em;
   list-style: none;
 
-  li::before {
+  & li::before {
     content: attr(data-icon);
     padding-right: 5px;
   }
