+++ src/components/RecipeView/Steps.module.css
@@ -1,6 +1,3 @@
-@import '../../global-styles/base.module.scss';
-@import '../../global-styles/globals.scss';
-
 .stepsContainer {
   margin-bottom: 20px;
 }
@@ -15,19 +12,19 @@
 
 .stepContainer {
   display: flex;
-  border-bottom: 1px dashed $lightBlue;
+  border-bottom: 1px dashed var(--light-blue);
 
   &:first-child {
-    border-top: 1px dashed $lightBlue;
+    border-top: 1px dashed var(--light-blue);
   }
 }
 
 .stepNumber {
   align-self: center;
-  flex: 0 0 30px; // ensures container width 30px
+  flex: 0 0 30px;
   font-size: 14px;
   font-weight: bold;
-  background-color: $lightBlue;
+  background-color: var(--light-blue);
   color: white;
   border-radius: 50%;
   margin-right: 10px;
@@ -42,7 +39,7 @@
 }
 
 .headerStepText {
-  color: $purple;
+  color: var(--purple);
   text-transform: uppercase;
   font-weight: 700;
   padding: 10px 0;
@@ -50,12 +47,12 @@
 
 .sectionHeader {
   font-weight: bold;
-  color: $purple;
+  color: var(--purple);
 }
 
 .superscript {
   font-size: 0.5em;
   font-weight: 600;
   vertical-align: super;
-  color: $purple;
+  color: var(--purple);
 }
