+++ src/components/SiteFrame/Header/Header.module.css
@@ -1,13 +1,10 @@
-@import '../../../global-styles/base.module.scss';
-@import '../../../global-styles/globals.scss';
-
 .headerContainer {
   position: fixed;
   top: 0;
   left: 0;
   right: 0;
   height: 50px;
-  background: $teal;
+  background: var(--teal);
   margin: 0;
   display: flex;
   align-items: center;
@@ -75,7 +72,7 @@
   display: flex;
   justify-content: flex-end;
 
-  .linkContainer {
+  & .linkContainer {
     text-align: right;
   }
 }
@@ -97,7 +94,7 @@
   right: 0;
   width: 160px;
   transition: opacity 0.5s ease;
-  background: $teal;
+  background: var(--teal);
   z-index: 1;
 }
