+++ src/components/base/Error.module.css
@@ -2,9 +2,11 @@
   text-align: center;
   font-size: 60px;
 }
+
 .icon {
   color: red;
 }
+
 .errorText {
   text-align: center;
 }
