+++ src/components/base/Error.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSadTear } from '@fortawesome/free-solid-svg-icons';
-import classes from './Error.module.scss';
+import classes from './Error.module.css';
 
 type Props = {
   errorText?: string;
