--- src/components/base/Loading.module.scss
@@ -1,61 +1,0 @@
-@mixin food($unique-identifier) {
-  $duration: 5;
-  @keyframes food-#{$unique-identifier} {
-    from {
-      transform: translateX(200%);
-      opacity: 1;
-    }
-    8% {
-      transform: translateX(0);
-    }
-    17% {
-      transform: translateX(0);
-      opacity: 1;
-    }
-    25% {
-      transform: translateX(-200%);
-      opacity: 0;
-    }
-  }
-  animation: food-#{$unique-identifier} $duration + s $unique-identifier + s ease-in-out infinite;
-}
-
-.loadingContainer {
-  text-align: center;
-  font-size: 45px;
-}
-
-.loadingText {
-  font-size: 18px;
-  margin-top: 20px;
-}
-
-.pizza {
-  color: #f49033;
-  opacity: 0;
-  @include food(0);
-}
-
-.hamburger {
-  color: #0081a7;
-  opacity: 0;
-  @include food(1);
-}
-
-.drumstick {
-  color: #00b959;
-  opacity: 0;
-  @include food(2);
-}
-
-.icecream {
-  color: #c836d5;
-  opacity: 0;
-  @include food(3);
-}
-
-.apple {
-  color: #f07167;
-  opacity: 0;
-  @include food(4);
-}
