+++ src/components/base/RecipeList/RecipeList.module.css
@@ -1,5 +1,3 @@
-@import '../../../global-styles/base.module.scss';
-
 .recipeListContainer {
   display: grid;
   grid-gap: 15px;
@@ -27,6 +25,7 @@
 }
 
 .recipeDetails {
+  flex: 1;
   padding: 5px 10px;
 
   &.red {
@@ -101,7 +100,6 @@
   min-height: 80px;
   color: white;
   fill: white;
-
   display: flex;
   justify-content: center;
   align-items: center;
@@ -133,10 +131,6 @@
   &.black {
     background-color: #000000;
   }
-}
-
-.recipeDetails {
-  flex: 1;
 }
 
 .tags {
