+++ src/components/base/RecipeList/RecipeList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Recipe as RecipeType } from '../../../types/recipe.types';
 import RecipeListItem from './RecipeListItem';
-import classes from './RecipeList.module.scss';
+import classes from './RecipeList.module.css';
 
 type Props = {
   recipes: RecipeType[];
