+++ src/global-styles/core.css
@@ -2,6 +2,12 @@
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@400;500;700&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Signika:wght@400;500;600;700&display=swap');
 
+:root {
+  --teal: #6796a7;
+  --purple: #6d2bb3;
+  --light-blue: #40b7f5;
+}
+
 body,
 html {
   margin: 0;
