--- src/global-styles/globals.scss
@@ -1,3 +1,0 @@
-$lightBlue: #40b7f5;
-$purple: #6d2bb3;
-$teal: #6796a7;
